requirements.txt
boto3
opentelemetry-exporter-otlp
opentelemetry-instrumentation-aws-lambda
opentelemetry-sdk
requests
